historical_odds/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.4.1
schedule>=1.2.0
pytz>=2024.1
psutil>=5.9.0
tqdm>=4.66.0
Flask>=3.0.0